// src/dorc-web/src/apis/dorc-api/models/PagedDataSorting.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface PagedDataSorting
 */
export interface PagedDataSorting {
    /**
     * @type {string}
     * @memberof PagedDataSorting
     */
    Path?: string | null;
    /**
     * @type {string}
     * @memberof PagedDataSorting
     */
    Direction?: string | null;
}