// src/dorc-web/src/apis/dorc-api/models/GetPropertyValuesAuditListResponseDto.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { PropertyValueAuditApiModel } from './';

/**
 * @export
 * @interface GetPropertyValuesAuditListResponseDto
 */
export interface GetPropertyValuesAuditListResponseDto {
  /**
   * @type {number}
   * @memberof GetPropertyValuesAuditListResponseDto
   */
  CurrentPage?: number;
  /**
   * @type {number}
   * @memberof GetPropertyValuesAuditListResponseDto
   */
  TotalItems?: number;
  /**
   * @type {number}
   * @memberof GetPropertyValuesAuditListResponseDto
   */
  TotalPages?: number;
  /**
   * @type {Array<PropertyValueAuditApiModel>}
   * @memberof GetPropertyValuesAuditListResponseDto
   */
  Items?: Array<PropertyValueAuditApiModel> | null;
}