// src/dorc-web/src/apis/dorc-api/models/PropertyValueAuditApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface PropertyValueAuditApiModel
 */
export interface PropertyValueAuditApiModel {
    /**
     * @type {number}
     * @memberof PropertyValueAuditApiModel
     */
    Id?: number;
    /**
     * @type {number}
     * @memberof PropertyValueAuditApiModel
     */
    PropertyId?: number | null;
    /**
     * @type {number}
     * @memberof PropertyValueAuditApiModel
     */
    PropertyValueId?: number | null;
    /**
     * @type {string}
     * @memberof PropertyValueAuditApiModel
     */
    PropertyName?: string | null;
    /**
     * @type {string}
     * @memberof PropertyValueAuditApiModel
     */
    EnvironmentName?: string | null;
    /**
     * @type {string}
     * @memberof PropertyValueAuditApiModel
     */
    FromValue?: string | null;
    /**
     * @type {string}
     * @memberof PropertyValueAuditApiModel
     */
    ToValue?: string | null;
    /**
     * @type {string}
     * @memberof PropertyValueAuditApiModel
     */
    UpdatedBy?: string | null;
    /**
     * @type {string}
     * @memberof PropertyValueAuditApiModel
     */
    UpdatedDate?: string;
    /**
     * @type {string}
     * @memberof PropertyValueAuditApiModel
     */
    Type?: string | null;
    /**
     * @type {boolean}
     * @memberof PropertyValueAuditApiModel
     */
    CanEdit?: boolean;
    /**
     * @type {boolean}
     * @memberof PropertyValueAuditApiModel
     */
    CanDelete?: boolean;
    /**
     * @type {boolean}
     * @memberof PropertyValueAuditApiModel
     */
    IsEnabled?: boolean;
}